# Refactor for better code readability