… .vscode/settings.json
… README.md
… dist/jsonpatcherproxy.js