I have made several improvements that were discussed with @tomalec over Slack. These changes are intended to increase code readability, which is needed before we can do further work.