@warpech 2019-06-04
### Compatibility

I have searched Palindrom and Starcounter GitHub orgs, and could not find any usage of the APIs that I've changed.

### Proposed more changes

We could change all properties to begin with `_`, because they are intended for private use. Private methods do begin with `_`.

### Benchmarks

#### Before this PR

```
jsonpatcherproxy generate operation
 Ops/sec: 136838 ±2.76% Ran 8652 times in 0.063 seconds.
jsonpatch generate operation
 Ops/sec: 109856 ±6.94% Ran 6494 times in 0.059 seconds.
jsonpatcherproxy mutation - huge object
 Ops/sec: 1283 ±0.26% Ran 67 times in 0.052 seconds.
jsonpatch mutation - huge object
 Ops/sec: 887 ±1.73% Ran 46 times in 0.052 seconds.
jsonpatcherproxy generate operation - huge object
 Ops/sec: 1084 ±0.99% Ran 58 times in 0.054 seconds.
jsonpatch generate operation - huge object
 Ops/sec: 860 ±1.65% Ran 47 times in 0.055 seconds.
PROXIFY big object
 Ops/sec: 2536 ±0.40% Ran 131 times in 0.052 seconds.
DIRTY-OBSERVE big object
 Ops/sec: 1746 ±1.18% Ran 94 times in 0.054 seconds.
```

#### After this PR

```
jsonpatcherproxy generate operation
 Ops/sec: 145558 ±2.37% Ran 9371 times in 0.064 seconds.
jsonpatch generate operation
 Ops/sec: 107443 ±5.93% Ran 6450 times in 0.060 seconds.
jsonpatcherproxy mutation - huge object
 Ops/sec: 923 ±1.95% Ran 52 times in 0.056 seconds.
jsonpatch mutation - huge object
 Ops/sec: 878 ±1.48% Ran 46 times in 0.052 seconds.
jsonpatcherproxy generate operation - huge object
 Ops/sec: 826 ±1.97% Ran 48 times in 0.058 seconds.
jsonpatch generate operation - huge object
 Ops/sec: 874 ±1.65% Ran 48 times in 0.055 seconds.
PROXIFY big object
 Ops/sec: 1845 ±2.16% Ran 103 times in 0.056 seconds.
DIRTY-OBSERVE big object
 Ops/sec: 1738 ±1.10% Ran 94 times in 0.054 seconds.
```