@warpech 2019-06-05
To verify that the TS type definitions are correct, I created a file locally `src/test.ts` with the following content and checked that it compiles:

```ts
import JSONPatcherProxy from '../';

var obj = {};
var instance = new JSONPatcherProxy(obj);
instance.generate();
instance.observe(true);
```