@alshakero 2019-06-05
It seems `.vscode/settings.json` was added unintentionally.