@warpech 2019-06-05
I actually did that intentionally, thinking that a common project config might be useful. Do you think it is better to remove it? @tomalec?