@tomalec 2019-06-05
I'm not sure. Usually, I'm not a fan of bloating project repo with IDE configs, but If think it's better to keep it, I'm good.